[Etc] animationWithCompletion 제거 및 viewModel 내부로 변경
[description]

[issue-number]
