+++ Sources/BezierSwift/MasterComponent/BezierDialog/BezierDialogViewModel.swift
@@ -7,6 +7,10 @@
 
 import SwiftUI
 
+private enum Constant {
+  static let animationDuration = 0.35.f
+}
+
 final class BezierDialogViewModel: ObservableObject {
   @Published var item: BezierDialogItem?
   
@@ -19,10 +23,17 @@ final class BezierDialogViewModel: ObservableObject {
       return
     }
     
-    animationWithCompletion(.easeInOut) { [weak self] in
+    guard self.item.isNotNil else {
+      withAnimation(.easeInOut(duration: Constant.animationDuration)) { [weak self] in
+        self?.item = item
+      }
+      return
+    }
+    
+    animationWithCompletion(.easeInOut(duration: Constant.animationDuration)) { [weak self] in
       self?.clearItem()
     } completion: { [weak self] in
-      withAnimation(.easeInOut) { [weak self] in
+      withAnimation(.easeInOut(duration: Constant.animationDuration)) { [weak self] in
         self?.item = item
       }
     }
@@ -31,13 +42,13 @@ final class BezierDialogViewModel: ObservableObject {
   func dismiss(with id: UUID) {
     guard self.item?.id == id else { return }
     
-    withAnimation(.easeInOut) { [weak self] in
+    withAnimation(.easeInOut(duration: Constant.animationDuration)) { [weak self] in
       self?.clearItem()
     }
   }
   
   func dismiss() {
-    withAnimation(.easeInOut) { [weak self] in
+    withAnimation(.easeInOut(duration: Constant.animationDuration)) { [weak self] in
       self?.clearItem()
     }
   }
@@ -45,5 +56,21 @@ final class BezierDialogViewModel: ObservableObject {
   private func clearItem() {
     self.item?.onDismiss?()
     self.item = nil
+  }
+  
+  private func animationWithCompletion(
+    _ animation: Animation,
+    _ body: () throws -> Void,
+    completion: @escaping () -> Void
+  ) rethrows {
+    if #available(iOS 17.0, *) {
+      try withAnimation(animation, body, completion: completion)
+    } else {
+      try withAnimation(animation, body)
+      
+      DispatchQueue.main.asyncAfter(deadline: .now() + Constant.animationDuration) {
+        completion()
+      }
+    }
   }
 }
